P4C-619: Hash Action in Brig (#2262)
* Initial table layout for hash action tables

* Added an input xbar allocation for the hash action table

* Changed the allocation for memories, resource_estimation, and table placement to handle hash action

* Fixed some bugs in order for the hash action example to generate assembly/Context JSON.  Still have an issue with determining the hash functions

A couple small hash function fixes

Some small fixes to get some more hash action corner cases

* A couple small register level bugs for set up

* Reverse in the input xbar and a rewrite of add_hash_functions for hash action tables

* Some small fixes to the constraints of hash action as well as better error messages on limitations of hash

* Table format bug exposed by the table placement changes

* Folded sort back into resource_estimate code as well as correctly determined when a table is considered hash action vs match with no key

* Added some comments to the hash action assembly functions

* Comments, error message cleanup

* Fixed warnings and an expand mis-calculation

* Small fixes to hash action assumptions/table format

* Removed some LOG messages

* A table format bug

Actually fixed the bug

* Change P4 in order to correctly validate the error messages

* Added some PR fixes

* Finalized some error messages

* Added another XFAIL

* Addressed Calin's PR comments

+++ bf-asm/counter.cpp
@@ -131,7 +131,7 @@ unsigned CounterTable::determine_shiftcount(Table::Call &call, int group, unsign
     if (call.args[0].name() && strcmp(call.args[0].name(), "$DIRECT") == 0) {
         return direct_shiftcount() + tcam_shift;
     } else if (call.args[0].field()) {
-        assert(call.args[0].field()->by_group[group]->bit(0)/128 == word);
+        assert(unsigned(call.args[0].field()->by_group[group]->bit(0)/128) == word);
         return call.args[0].field()->by_group[group]->bit(0)%128 + indirect_shiftcount();
     } else if (call.args[1].field()) {
         return call.args[1].field()->bit(0) + STAT_ADDRESS_ZERO_PAD;
